ref(node-core): Move node fetch instrumentation into node-core
Move the `undici-instrumentation` (spans + breadcrumbs + trace propagation) and the vendored
`node-fetch` types from the `node` package into `node-core`, and expose `instrumentUndici` /
`NodeFetchOptions` from `@sentry/node-core`.

`node-core`'s `nativeNodeFetchIntegration` now instruments via `instrumentUndici` directly
(spans default to `true`). The `node` package keeps a thin `nativeNodeFetchIntegration` wrapper
(`node-fetch.ts`) that derives the `spans` default from the client options
(`skipOpenTelemetrySetup` / `hasSpansEnabled`) before delegating to `instrumentUndici`.

`SentryNodeFetchInstrumentation` is now unused internally and is marked `@deprecated`; it stays
exported for backwards compatibility and will be removed in a future major version.
