+++ packages/node-core/src/index.ts
@@ -11,11 +11,13 @@ export {
   type SentryHttpInstrumentationOptions,
 } from './integrations/http/SentryHttpInstrumentation';
 export { nativeNodeFetchIntegration } from './integrations/node-fetch';
+export type { NodeFetchOptions } from './integrations/node-fetch/types';
+export { instrumentUndici } from './integrations/node-fetch/undici-instrumentation';
 export {
+  // oxlint-disable-next-line typescript/no-deprecated
   SentryNodeFetchInstrumentation,
   type SentryNodeFetchInstrumentationOptions,
 } from './integrations/node-fetch/SentryNodeFetchInstrumentation';
-export { addFetchRequestBreadcrumb, addTracePropagationHeadersToFetchRequest } from './utils/outgoingFetchRequest';
 
 export { SentryContextManager } from './otel/contextManager';
 export { setupOpenTelemetryLogger } from './otel/logger';
