+++ packages/node-core/src/integrations/node-fetch/types.ts
@@ -1,23 +1,16 @@
 /*
  * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * NOTICE from the Sentry authors:
+ * - Vendored from: https://github.com/open-telemetry/opentelemetry-js-contrib/tree/ed97091c9890dd18e52759f2ea98e9d7593b3ae4/packages/instrumentation-undici
+ * - Upstream version: @opentelemetry/instrumentation-undici@0.24.0
+ * - Tracking issue: https://github.com/getsentry/sentry-javascript/issues/20165
+ * - Dropped the `@opentelemetry/instrumentation` `InstrumentationConfig` base (its only field used
+ *   here, `enabled`, is unused by the Sentry integration)
  */
 
-/**
- * Aligned with upstream Undici request shape; see `packages/node/.../node-fetch/vendored/types.ts`
- * (vendored from `@opentelemetry/instrumentation-undici`).
- */
+import type { Span } from '@sentry/core';
 
 export interface UndiciRequest {
   origin: string;
@@ -25,7 +18,8 @@ export interface UndiciRequest {
   path: string;
   /**
    * Serialized string of headers in the form `name: value\r\n` for v5
-   * Array of strings `[key1, value1, ...]` for v6 (values may be `string | string[]`)
+   * Array of strings `[key1, value1, key2, value2]`, where values are
+   * `string | string[]` for v6
    */
   headers: string | (string | string[])[];
   /**
@@ -38,11 +32,89 @@ export interface UndiciRequest {
   idempotent: boolean;
   contentLength: number | null;
   contentType: string | null;
-  body: unknown;
+  // oxlint-disable-next-line typescript/no-explicit-any
+  body: any;
 }
 
 export interface UndiciResponse {
   headers: Buffer[];
   statusCode: number;
   statusText: string;
+}
+
+export interface RequestHookFunction<T = UndiciRequest> {
+  (span: Span, request: T): void;
+}
+
+export interface ResponseHookFunction<RequestType = UndiciRequest, ResponseType = UndiciResponse> {
+  (span: Span, info: { request: RequestType; response: ResponseType }): void;
+}
+
+export interface RequestMessage {
+  request: UndiciRequest;
+}
+
+export interface RequestHeadersMessage {
+  request: UndiciRequest;
+  // oxlint-disable-next-line typescript/no-explicit-any
+  socket: any;
+}
+
+export interface ResponseHeadersMessage {
+  request: UndiciRequest;
+  response: UndiciResponse;
+}
+
+export interface RequestTrailersMessage {
+  request: UndiciRequest;
+  response: UndiciResponse;
+}
+
+export interface RequestErrorMessage {
+  request: UndiciRequest;
+  error: Error;
+}
+
+// This package will instrument HTTP requests made through `undici` or  `fetch` global API
+// so it seems logical to have similar options than the HTTP instrumentation
+export interface UndiciInstrumentationConfig<RequestType = UndiciRequest, ResponseType = UndiciResponse> {
+  /**
+   * Do not capture spans or breadcrumbs for outgoing fetch requests to URLs where the given callback returns `true`.
+   * This controls both span & breadcrumb creation - spans will be non recording if tracing is disabled.
+   */
+  ignoreOutgoingRequests?: (url: string) => boolean;
+  /** Function for adding custom attributes before request is handled */
+  requestHook?: RequestHookFunction<RequestType>;
+  /** Function called once response headers have been received */
+  responseHook?: ResponseHookFunction<RequestType, ResponseType>;
+  /** Map the following HTTP headers to span attributes. */
+  headersToSpanAttributes?: {
+    requestHeaders?: string[];
+    responseHeaders?: string[];
+  };
+}
+
+export interface NodeFetchOptions extends UndiciInstrumentationConfig {
+  /**
+   * Whether breadcrumbs should be recorded for requests.
+   *
+   * @default `true`
+   */
+  breadcrumbs?: boolean;
+
+  /**
+   * If set to false, do not emit any spans.
+   * Breadcrumbs and trace propagation for outgoing fetch requests are still applied.
+   *
+   * If `skipOpenTelemetrySetup: true` is configured, this defaults to `false`, otherwise it defaults to `true`.
+   */
+  spans?: boolean;
+
+  /**
+   * This option only has an effect when `spans` is set to false. When spans are enabled, you cannot disable trace propagation here.
+   * Instead, configure `tracePropagationTargets` in the client options.
+   *
+   * @default `true`
+   */
+  tracePropagation?: boolean;
 }
